test: move some BrowserWindow specs to the main runner (#18551)
